+++ src-tauri/src/lib.rs
@@ -61,6 +61,8 @@ pub fn run() {
   tauri::Builder::default()
     .plugin(tauri_plugin_shell::init())
     .plugin(tauri_plugin_dialog::init())
+    .plugin(tauri_plugin_updater::Builder::new().build())
+    .plugin(tauri_plugin_process::init())
     .plugin(tauri_plugin_global_shortcut::Builder::new()
       .with_handler(|app, _shortcut, event| desktop::handle_global_shortcut(app, event.state()))
       .build())